+++ src/lane_keep_assist/run_metrics.py
@@ -1,4 +1,4 @@
-from lane_test import lane_detect_test
+from testing.PythonTests.lane_keep_assist import lane_detect_test
 
 
 def main():
