+++ testing/PythonTests/lane_keep_assist/take_picture.py
